+++ tests/integration/components/contributor-list/component-test.js
@@ -1,14 +1,12 @@
 import EmberObject from '@ember/object';
-import EmberService from '@ember/service';
 import { moduleForComponent, test } from 'ember-qunit';
 import hbs from 'htmlbars-inline-precompile';
 
 
-// Stub store service
-const storeStub = EmberService.extend({
-    init() {
-        this._super(...arguments);
-        this.contributorList = {
+moduleForComponent('contributor-list', 'Integration | Component | contributor-list', {
+    integration: true,
+    beforeEach() {
+        const contributorList = () => ({
             data: [
                 EmberObject.create({ // Records to be returned by queryHasMany
                     users: { givenName: 'Brian', familyName: 'Nosek', profileURL: 'https://osf.io/12345' },
@@ -23,24 +21,20 @@ const storeStub = EmberService.extend({
             toArray() {
                 return this.data;
             },
-        };
-    },
-    queryHasMany () {
-        return this.get('contributorList');
-    },
-});
+        });
 
-moduleForComponent('contributor-list', 'Integration | Component | contributor-list', {
-    integration: true,
-    beforeEach() {
-        this.register('service:store', storeStub);
-        this.inject.service('store', { as: 'store' });
+        const node = EmberObject.create({
+            content: {
+                queryHasMany: contributorList,
+            },
+        });
+
+        this.set('node', node);
     },
 });
 
 
 test('it renders contributor-list', function(assert) {
-    this.set('node', {});
     this.render(hbs`{{contributor-list contributors=null node=node}}`);
     assert.ok(this.$('ul').length);
     assert.equal(